Add groups controller || Fix schema sql

+++ db/migrate/20220523174651_create_agreements.rb
@@ -2,7 +2,7 @@ class CreateAgreements < ActiveRecord::Migration[7.0]
   def change
     create_table :agreements do |t|
       t.string :name
-      t.float :amount, default: 0
+      t.decimal :amount, default: 0
       t.integer :author_id
 
       t.timestamps
